[#1] refactor: gin server interface 추상화
